requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.6.4
python-multipart==0.0.9
jinja2==3.1.3
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
